Created a new script called PageNavigator.tsx to store all page-navigation functions. Adapted all other scripts with this change as well.
