+++ linkedup/app/main/page.tsx
@@ -1,7 +1,5 @@
 'use client';
 
-import Image from 'next/image';
-import { useRouter } from 'next/navigation';
 import { useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 import CalendarTimeline from '@/components/CalendarTimeline';
@@ -9,6 +7,7 @@ import ActivityPanel from '@/components/ActivityPanel';
 import GoToTopButton from '@/components/GoToTopButton';
 import type { Activity } from '@/schemas/ActivityRelated';
 import { ActivityTable } from '@/schemas/ActivityRelated';
+import { GoToProfilePage, GoToActivityHistoryPage, GoToActivityCreationPage } from '@/components/PageNavigator';
 
 // Mock data
 const myActivityLst: Activity[] = [
@@ -65,38 +64,6 @@ const activityFeedsList: Activity[] = [
      maxAttendees: 5, participants: []
     },
 ]
-
-// Transits to another page without using the 'Link' attribute
-function GoToActivityCreationPage() {
-    const router = useRouter();
-
-    return (
-        <button onClick={() => router.push('/activity-creation')} 
-                className='mt-4 block w-full rounded-md bg-sky-400 py-2 text-center font-semibold text-white hover:bg-sky-500 transition cursor-pointer'>
-            Post an Activity
-        </button>
-    );
-}
-function GoToActivityHistoryPage() {
-    const router = useRouter();
-
-    return (
-        <button onClick={() => router.push('/activity-history')} 
-                className='mt-4 block w-full rounded-md bg-sky-400 py-2 text-center font-semibold text-white ≈ transition cursor-pointer'>
-            Activity History
-        </button>
-    );
-}
-function GoToProfilePage() {
-    const router = useRouter();
-
-    return (
-        <div onClick={() => router.push('/profile')} 
-             className='h-12 w-12 rounded-full overflow-hidden border cursor-pointer'>
-            <Image src='/lemon_drink.jpeg' alt='User Profile Picture' width={48} height={48} className='object-cover'/>
-        </div>
-    )
-}
 
 export default function MainPage() {
     const username = 'user123';
